Move THIR printing to rustc_mir_build.
